(refactor) Refactor tests.
