+++ test/tests/barChartTest.js
