+++ test/tests/coreTest.js
