+++ test/tests/scatterChartTest.js
